move run-liquidation-bot
